Hackathon-main/backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
groq
python-dotenv
pydantic
